Add default protocol, basic validation
